Refactor TUI state helpers
